♻️ Add Divider to Popover.Header
