+++ packages/eds-core-react/src/components/Popover/Popover.stories.tsx
@@ -309,14 +309,15 @@ export const WithCloseButton: Story<PopoverProps> = () => {
         >
           <Popover.Header>
             <Popover.Title>Title</Popover.Title>
-            <StyledCloseButton
+            <Button
               variant="ghost_icon"
               aria-label="Close popover"
+              title="Close popover"
               data-testid="popover-close"
               onClick={closePopover}
             >
               <Icon name="close" data={close} size={24} />
-            </StyledCloseButton>
+            </Button>
           </Popover.Header>
           <Popover.Content>
             <Typography variant="body_short">Content</Typography>
