+++ packages/eds-core-react/src/components/Popover/Popover.tsx
@@ -182,7 +182,6 @@ export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
               <path d="M0.504838 4.86885C-0.168399 4.48524 -0.168399 3.51476 0.504838 3.13115L6 8.59227e-08L6 8L0.504838 4.86885Z" />
             </PopoverArrow>
           </ArrowWrapper>
-
           {children}
         </PopoverPaper>
       </ThemeProvider>
