+++ packages/eds-core-react/src/components/Popover/PopoverHeader.tsx
@@ -1,11 +1,23 @@
 import { forwardRef, HTMLAttributes } from 'react'
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
+import { Divider } from '../Divider'
 
 export type PopoverHeaderProps = HTMLAttributes<HTMLDivElement>
 
 const StyledPopoverHeader = styled.div`
-  display: block;
+  display: grid;
+  grid-auto-flow: column;
 `
+
+const StyledDivider = styled(Divider)(({ theme }) => {
+  return css`
+    margin-left: -${theme.spacings.left};
+    margin-right: -${theme.spacings.right};
+    margin-bottom: 0;
+    width: auto;
+    max-width: ${theme.maxWidth};
+  `
+})
 
 export const PopoverHeader = forwardRef<HTMLDivElement, PopoverHeaderProps>(
   function PopoverHeader({ children, ...rest }, ref) {
@@ -14,6 +26,11 @@ export const PopoverHeader = forwardRef<HTMLDivElement, PopoverHeaderProps>(
       ref,
     }
 
-    return <StyledPopoverHeader {...props}>{children}</StyledPopoverHeader>
+    return (
+      <div {...props}>
+        <StyledPopoverHeader>{children}</StyledPopoverHeader>
+        <StyledDivider variant="small" />
+      </div>
+    )
   },
 )
