+++ packages/eds-core-react/src/components/Popover/PopoverTitle.tsx
@@ -1,6 +1,5 @@
 import { forwardRef, HTMLAttributes } from 'react'
 import styled, { css } from 'styled-components'
-import { Divider } from '../Divider'
 
 const StyledPopoverTitle = styled.div(({ theme }) => {
   return css`
@@ -8,16 +7,6 @@ const StyledPopoverTitle = styled.div(({ theme }) => {
     max-width: calc(${theme.maxWidth} - ${theme.entities.closeButton.width});
     overflow: hidden;
     margin-top: -${theme.entities.title.spacings.top};
-  `
-})
-
-const StyledDivider = styled(Divider)(({ theme }) => {
-  return css`
-    margin-left: -${theme.spacings.left};
-    margin-right: -${theme.spacings.right};
-    margin-bottom: 0;
-    width: auto;
-    max-width: ${theme.maxWidth};
   `
 })
 
@@ -30,11 +19,6 @@ export const PopoverTitle = forwardRef<HTMLDivElement, PopoverTitleProps>(
       ref,
     }
 
-    return (
-      <div {...props}>
-        <StyledPopoverTitle>{children}</StyledPopoverTitle>
-        <StyledDivider variant="small" />
-      </div>
-    )
+    return <StyledPopoverTitle {...props}>{children}</StyledPopoverTitle>
   },
 )
